home page invoices front

+++ OAS.Application/Features/InvoiceFeatures/GetSellInvoices/GetSellInvoices_InvoiceDTO.cs
@@ -7,5 +7,5 @@ using System.Threading.Tasks;
 namespace OAS.Application.Features.InvoiceFeatures.GetSellInvoices
 {
     public sealed record GetSellInvoices_InvoiceDTO(int RowNumber,Guid InvoiceId , string InvoiceCode , DateTime InvoiceDate,string InvoiceDateString,string VehicleName , 
-        string CustomerName,long  InvoiceTotalPrice , long InvoicePaidAmount);    
+        string CustomerName,float  InvoiceTotalPrice , float InvoicePaidAmount);    
 }
